+++ src/ai/backend/manager/repositories/metric/repositories.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
 from typing import Self
 
-from ai.backend.manager.repositories.image.repositories import RepositoryArgs
 from ai.backend.manager.repositories.metric.repository import MetricRepository
+from ai.backend.manager.repositories.types import RepositoryArgs
 
 
 @dataclass
