+++ src/ai/backend/manager/repositories/project_resource_policy/repositories.py
@@ -1,10 +1,10 @@
 from dataclasses import dataclass
 from typing import Self
 
-from ai.backend.manager.repositories.image.repositories import RepositoryArgs
 from ai.backend.manager.repositories.project_resource_policy.repository import (
     ProjectResourcePolicyRepository,
 )
+from ai.backend.manager.repositories.types import RepositoryArgs
 
 
 @dataclass
